requirements.txt
pillow~=11.0.0
numpy~=2.2.0
Flask~=3.1.0
gunicorn~=22.0.0